virttest.qemu_vm: Adjust reboot go_down_method
Since 3c2f6241a when the VM's monitor is QMP it always uses it to check
for reboot even though other methods might already define their check
methods as it claims to be the most reliable method. Let's keep this by
default, but allow disabling this behavior and allowing to use any
available method instead. One example is the "with_reboot" migration
test that constantly changes the VM object, which results to missed QMP
events.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>