virttest: Add support to skip VM state verification on login
In some cases the VM status might be unreliable/unobtainable (for
example when rebooting, stopping or suspending the VM) and the extra
"verify_alive" check leads to a false-negative.

Use this "status_check=False" in "VM.reboot" to prevent false-negatives
when running migrate.with_reboot tests that utilize the "VM.reboot"
together with "VM.migrate".

Signed-off-by: Xu Han <xuhan@redhat.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>